requirements.txt
Flask==3.0.3
flask-login==0.6.3
pyotp==2.9.0
qrcode==7.4.2
pillow==10.4.0
Werkzeug==3.0.3

cryptography~=46.0.3
python-dotenv~=1.2.1